requirements.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
requests>=2.31.0
beautifulsoup4>=4.14.0
# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
